requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
langchain
langchain-google-genai
langgraph
google-generativeai
markdown
langchain-community
razorpay